requirements.txt
streamlit==1.22.0
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
xgboost==1.4.2
joblib==1.0.1
groq==0.3.2
python-dotenv==1.0.0